# Optimize shfit_gram and shfit_window functions