Optimize shfit_gram and shfit_window functions by just increasing the pointer, if the buffer is going to exhausted,
left shift the whole size of grams/windows to leftmost once together.

I generate a 109M benchmark data, On my 2.71 GHz Intel Core i7 8559U CPU, 32 GB 2400 MHz DDR4 computer:

1. Just optimize shift_gram function, the performance was improved  from 9.5s  to 8.5s. (3 times running for average)
2. add optimizing of shift_window function, the performance was improved  from 8.5s  to 7.6s.
Improved performance by about 20%，but At the cost of increased code complexity a bit. Do you think it's worthy? please let me know.

I am not skilled C-programer, if there is any problem, feel free to point out.