@scanoss-qg 2021-12-10
Hello again Alec-z:  Thanks (one more time)  for your valuable contribution. We have been doing a deeper analsys and testing your proposal. Tests shown that your improvement really speeds up the process a about 20% (in particular for small files).  The complexity grows a bit, but with a further descriptive text of the optimizatuin, I think it would be clear.